+++ tests/test_embedded_js.py
@@ -10,12 +10,14 @@ div
 p
 """
 
-expected_output = """<div>
+expected_output = """\
+<div>
 <script type='javascript'>
     console.log('hello world');
     if (true) {
         alert('hello world');
     }
+
 </script></div><p></p>"""
 
 def test_embedded_js():